Fix LegalizeJSInterface with RefFuncs (#3749)
This code used to remove functions it no longer thinks are needed. That is,
if it adds a legalized version of an import, it would remove the illegal
one which is no longer needed. To avoid removing an illegal import that
is still used it checked for ref.func appearances.

But this was bad in two ways:

    We need to legalize the ref.funcs too. We can't call an illegal import
    in any way, not using a direct call, indirect call, or call by reference of
    a ref.func.

    It's silly to remove unneeded functions here. We have a pass for that.

This removes the removal of functions, and adds proper updating of
ref.calls, which means to call the stub function that looks like the
original import, but that calls the legalized one and connects things
up properly, exactly the same way as other calls.

Also remove code that checked if we were in the stub/thunk and to
not replace the call there. That code is not needed: no one will ever
call the illegal import, so we do not need to be careful about
preserving such calls.